+++ content/components/hidden.mdx
@@ -1,7 +1,7 @@
 ---
-title: Filter list
-description: The FilterList component is a menu with filter options that filter the main content of the page.
-reactId: filter_list
+title: Hidden
+description: Use Hidden to responsively hide or show content in narrow, regular and wide viewports.
+reactId: hidden
 ---
 
 import ComponentLayout from '~/src/layouts/component-layout'
@@ -19,4 +19,4 @@ import {AccessibilityLink} from '~/src/components/accessibility-link'
 
 ### Known accessibility issues (GitHub staff only)
 
-<AccessibilityLink label="FilterList"/>
+<AccessibilityLink label="Hidden"/>
