+++ content/deprecated-components/filtered-search.mdx
@@ -19,4 +19,4 @@ import {AccessibilityLink} from '~/src/components/accessibility-link'
 
 ### Known accessibility issues (GitHub staff only)
 
-<AccessibilityLink label="CloseButton"/>
+<AccessibilityLink label="FilteredSearch"/>
