+++ content/deprecated-components/markdown-viewer.mdx
@@ -1,7 +1,7 @@
 ---
-title: Filter list
-description: The FilterList component is a menu with filter options that filter the main content of the page.
-reactId: filter_list
+title: Markdown viewer
+description: Displays rendered Markdown and facilitates interaction.
+reactId: markdown_viewer
 ---
 
 import ComponentLayout from '~/src/layouts/component-layout'
@@ -19,4 +19,4 @@ import {AccessibilityLink} from '~/src/components/accessibility-link'
 
 ### Known accessibility issues (GitHub staff only)
 
-<AccessibilityLink label="FilterList"/>
+<AccessibilityLink label="MarkdownEditor"/>
